Project Title: Nashville Housing Data Cleaning
This project involves cleaning and preparing Nashville housing data in a database 
on an offline server using SQL Server Management Studio. The goal is to ensure
the data is accurate, consistent, and ready for analysis without exporting it 
from the secure server.

It involves handling missing values, standardizing formats, removing duplicates, 
correcting errors, and normalizing data.